+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 Author: Vadim Zaboturin <vadimzab@gmail.com>
 Date:   2022-04-11
 
+    #0.1.96
+    1. Скидка не применяется на страницах product и order если статус товара status == 'Outlet' || status == 'Discount' || status == 'Activity'
+    
+    
+Author: Vadim Zaboturin <vadimzab@gmail.com>
+Date:   2022-04-11
+
     #0.1.95
-    1. На странице с Отгрузки  добавил детали заказа.
+    1. На странице с Отгрузки  добавил детали заявки.
     2. Убрал фильтры на страницах Заказы Отгрузки Претензии.